requirements.txt
streamlit
google-cloud-bigquery
vertexai